IAM function definition moved to awsEventCallbacks file from main.js for readability
