+++ public/helperFunctions/awsEventCallbacks.js
@@ -1,4 +1,88 @@
+const awsHelperFunctions = require(__dirname + '/awsHelperFunctions'); 
+const awsParameters = require(__dirname + '/awsParameters'); 
+
+const EKS = require('aws-sdk/clients/eks');
+const IAM = require('aws-sdk/clients/iam');
+const CloudFormation = require('aws-sdk/clients/cloudformation');
+
+const fs = require('fs');
+const { spawn } = require('child_process');
+const fsp = require('fs').promises;
+
+
+//**.ENV Variables */
+const REGION = process.env.REGION;
+const SERVICEROLEARN = process.env.EKSSERVICEROLEARN;
+
+const SUBNETID1 = process.env.SUBNETID1;
+const SUBNETID2 = process.env.SUBNETID2;
+const SUBNETID3 = process.env.SUBNETID3;
+const SUBNETIDS = [SUBNETID1, SUBNETID2, SUBNETID3];
+const SECURITYGROUPIDS = process.env.SECURITYGROUPIDS;
+
+//** --------- INITIALIZE IMPORTS --------- 
+const iam = new IAM()
+const eks = new EKS({ region: REGION});
+const cloudformation = new CloudFormation({ region: REGION});
+
+
 const awsEventCallbacks = {};
+
+//** --------- CREATE AWS IAM ROLE + ATTACH POLICY DOCS --------------- **//
+awsEventCallbacks.createIAMRoleAndCreateFileAndAttachPolicyDocs = async (roleName, roleDescription, iamRolePolicyDoc) => {
+
+  const iamParams = awsParameters.createIAMRoleParam(roleName, roleDescription, iamRolePolicyDoc);
+
+  //Send IAM data to AWS via the iamParams object to create an IAM Role*/
+  try {
+    const role = await iam.createRole(iamParams).promise();
+
+    //Collect the relevant IAM data returned from AWS
+    const iamRoleDataFromForm = {
+      roleName: role.Role.RoleName,
+      createDate: role.Role.RoleId,
+      roleID: role.Role.RoleId,
+      arn: role.Role.Arn,
+      createDate: role.Role.CreateDate,
+      path: role.Role.Path
+    }
+
+    const stringifiedIamRoleDataFromForm = JSON.stringify(iamRoleDataFromForm);
+    
+    //Create file named for IAM Role and save in assets folder */
+    fsp.writeFile(__dirname + `/../sdkAssets/private/${roleName}.json`, stringifiedIamRoleDataFromForm);
+
+    //Send Cluster + Service Policies to AWS to attach to created IAM Role 
+    const clusterPolicyArn = 'arn:aws:iam::aws:policy/AmazonEKSClusterPolicy';
+    const servicePolicyArn = 'arn:aws:iam::aws:policy/AmazonEKSServicePolicy';
+
+    const clusterPolicy = {
+      RoleName: roleName,
+      PolicyArn: clusterPolicyArn
+    }
+
+    const servicePolicy = {
+      RoleName: roleName,
+      PolicyArn: servicePolicyArn
+    }
+
+    await iam.attachRolePolicy(clusterPolicy).promise();
+    await iam.attachRolePolicy(servicePolicy).promise();
+
+    //TODO: BRADON: PROMISE ALL INSTEAD
+  
+  } catch (err) {
+    console.log(err);
+  }
+
+};
+
+//** --------- CREATE AWS TECH STACK + SAVE RETURNED DATA IN FILE ----- **//
+awsEventCallbacks.createTechStackAndSaveReturnedDataInFile = async (roleName, roleDescription, iamRolePolicyDoc) => {
+
+
+
+
 
 //**--------- GENERATE AND SAVE CONFIG FILE ON USER COMPUTER ------------- **//
 
@@ -15,74 +99,7 @@ awsEventCallbacks.CREATE_CONFIG_FILE = (clusterName) => {
   const clusterArn = parsedclusterDataFileContents.cluster.arn;
   const certificateAuthorityData = parsedclusterDataFileContents.cluster.certificateAuthority.data;
 
-  const AWSClusterConfigFileData = {
-    "apiVersion": "v1",
-    "clusters": [
-        {
-            "cluster": {
-                "server": serverEndpoint,
-                "certificate-authority-data": certificateAuthorityData,
-            },
-            "name": "kubernetes"
-        },
-        {
-            "cluster": {
-                "certificate-authority-data": certificateAuthorityData,
-                "server": serverEndpoint
-            },
-            "name": clusterArn
-        }
-    ],
-    "contexts": [
-        {
-            "context": {
-                "cluster": "kubernetes",
-                "user": "aws"
-            },
-            "name": "aws"
-        },
-        {
-            "context": {
-                "cluster": clusterArn,
-                "user": clusterArn
-            },
-            "name": clusterArn
-        }
-    ],
-    "current-context": clusterArn,
-    "kind": "Config",
-    "preferences": {},
-    "users": [
-        {
-            "name": "aws",
-            "user": {
-                "exec": {
-                    "apiVersion": "client.authentication.k8s.io/v1alpha1",
-                    "command": "aws-iam-authenticator",
-                    "args": [
-                        "token",
-                        "-i",
-                        clusterName
-                    ]
-                }
-            }
-        },
-        {
-            "name": clusterArn,
-            "user": {
-                "exec": {
-                    "apiVersion": "client.authentication.k8s.io/v1alpha1",
-                    "args": [
-                        "token",
-                        "-i",
-                        clusterName
-                    ],
-                    "command": "aws-iam-authenticator"
-                }
-            }
-        }
-    ]
-  }
+  const AWSClusterConfigFileData = awsParameters.createConfigParam(clusterName, serverEndpoint, certificateAuthorityData, clusterArn);
   
   //** Format data from the AWSClusterConfigFileData object into YAML to save in user's filesystem 
   let stringifiedAWSClusterConfigFile = JSON.stringify(AWSClusterConfigFileData);
@@ -91,12 +108,11 @@ awsEventCallbacks.CREATE_CONFIG_FILE = (clusterName) => {
   let regexCharToRemove = /(['])+/g;
   let yamledAWSClusterConfigFileWithoutRegex = yamledAWSClusterConfigFile.replace(regexCharToRemove, "");
 
-  //TODO: .Kube directory, test if ~./kube directory exists, if not, mkdir -p ~/.kube
+  //check if user has a .kube file in their root directory, and if not, make one
+  awsHelperFunctions.checkFileSystemForDirectoryAndMkDir(kube);
 
-  
-
-  //** Save data in users .kube file
-  fs.writeFileSync('/Users/carolynharrold/.kube/config-carolyn-and-adrian-killer-cluster', yamledAWSClusterConfigFileWithoutRegex);
+  //Save file in users .kube file
+  fs.writeFileSync(`/Users/carolynharrold/.kube/{clusterName}`, yamledAWSClusterConfigFileWithoutRegex);
 
 };
